refactor(repo C8): extract context annotations to engine/repository/context-annotations.ts
Moved insertContext, deleteContext, deleteGlobalContexts, listPathContexts,
getContextForPath, getCollectionsWithoutContext, and getTopLevelPathsWithoutContext.
getContextForFile stays in repository.ts because it depends on parseVirtualPath
which has not yet been extracted.

Build: clean. Tests: 970/970 pass.
